moved content of components folder to ui
