+++ app/(pages)/filmer/page.tsx
@@ -1,5 +1,5 @@
 'use client';
-import MovieList from '@/app/components/filmer/MovieList';
+import MovieList from '@/app/ui/filmer/MovieList';
 import { Typography } from '@mui/material';
 
 export default function MovieListPage() {
