+++ app/(pages)/login/page.tsx
@@ -1,4 +1,4 @@
-import Login from '@/app/ui/login';
+import Login from '@/app/ui/login/login';
 
 export default function LoginPage() {
   return <Login></Login>;
