+++ app/(pages)/profil/page.tsx
@@ -1,4 +1,4 @@
-import UserProfile from '@/app/ui/userProfile';
+import UserProfile from '@/app/ui/login/userProfile';
 
 export default function ProfilePage() {
   return <UserProfile />;
