+++ app/(pages)/registrera/page.tsx
@@ -1,4 +1,4 @@
-import SignUp from '@/app/ui/signup';
+import SignUp from '@/app/ui/login/signup';
 
 export default function Registrera() {
   return <SignUp></SignUp>;
