+++ app/ui/landing/screeningsTableStartpage.tsx
@@ -12,7 +12,7 @@ import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import dayjs, { Dayjs } from 'dayjs';
 import 'dayjs/locale/sv';
-import { Screening } from '../lib/definitions';
+import { Screening } from '../../lib/definitions';
 import Link from 'next/link';
 
 export default function ScreeningsTableStartpage() {
